ajout create blog

+++ database/migrations/2017_02_09_112316_create_blogs_table.php
@@ -18,8 +18,8 @@ class CreateBlogsTable extends Migration
 			$table->string('picture');
 			$table->string('ville');
 			$table->string('pays');
-			$table->integer('longitude');
-			$table->integer('latitude');
+			$table->decimal('longitude');
+			$table->decimal('latitude');
             $table->longText('blog_text');
             $table->rememberToken();
             $table->timestamps();
